#576 refactoring of ap updating in aa
